requirements.txt
# Core dependencies
streamlit
python-dotenv
phidata
groq
requests
langchain
langchain-core
langchain-groq

# Search tools dependencies
duckduckgo-search
google-search-results
googlesearch-python